Week 2 Homework
This is my code for the week 2 homework assignment. When I added some of the elements that were in the design file, the landing page did not look like the sample single-column picture so I adjusted sizes, etc. to get it as close to the picture as possible. Thank you!