Brooke Homework - Lesson 13 Responsive Design
I had trouble getting the nav bar to flex without the header getting a white space underneath it. I tried to troubleshoot but it just kept getting worse. I used vh and vw units for the font changes. I know we didn't really do this in class but it was something i learned on Team Treehouse and seems to work for the most part. I tried the em and rem method and the v method just seemed to be more fluid. 
I got the blog to one column and did the aside underneath it. Let me know if you have questions. 
Thanks!